pyproject.toml
[tool.poetry]
    name        = "MSFS Mod Manager"
    version     = "2.0.0"
    description = "External mod manager for the new Microsoft Flight Simulator"
    authors     = ["Nathan Vaughn <nvaughn51@gmail.com>"]
    license     = "GPL-3.0-or-later"
    readme      = "README.md"
    homepage    = "https://github.com/NathanVaughn/msfs-mod-manager"
    repository  = "https://github.com/NathanVaughn/msfs-mod-manager"

[tool.poetry.dependencies]
    python  = ">=3.11,<3.12"
    loguru  = "^0.6.0"
    patool  = "^1.12"
    PySide6-Essentials = "^6.5.2" # 6.2.2 has a bug that makes our threads stop working

[tool.poetry.dev-dependencies]
    pre-commit  = "^3.2.0"
    pyinstaller = "^6.0"

[tool.pyright]
    typeCheckingMode           = "basic"
    venvPath                   = "."
    venv                       = ".venv"

[tool.ruff]
    target-version = "py311"
    ignore         = ["E501"]

[build-system]
    requires      = ["poetry-core>=1.0.0"]
    build-backend = "poetry.core.masonry.api"